Migrated SEI graph
