+++ dashboard/dashboard.py
@@ -9,7 +9,7 @@ import plotly.express as px
 from dash import dcc, html
 from nltk.corpus import stopwords
 
-from data import load_assignment_survey_data
+from data import load_assignment_survey_data, load_sei_data
 import callbacks
 
 # Constants
@@ -36,24 +36,6 @@ satisfaction_mapping = {
 likert_scale = ["Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"]
 likert_scale_alt = ["Poor", "Fair", "Satisfactory", "Very good", "Excellent"]
 satisfaction_colors = dict(zip(satisfaction_mapping.values(), px.colors.sequential.Viridis[::2]))
-
-def _semester_order(data: pd.DataFrame):
-  """
-  Returns a sorted list of semesters in the expected order 
-  (e.g., [Autumn 2018, Spring 2019, Autumn 2019, Spring 2020, ...]).
-  
-  It works by parsing the semester string and calculating a
-  sortable numeric value where the year is used unless
-  the semester is in the autumn, in which case the year + .5
-  is used. 
-
-  :param data: the DataFrame provided by the user with an assumed Semester column
-  :return: a list of sorted semesters
-  """
-  return sorted(
-    data["Semester"].unique(), 
-    key=lambda x: int(x.split()[1]) + (.5 if x.split()[0] == "Autumn" else 0)
-  )
 
 def create_value_fig(grade_data, assignment_survey_data, assignment, max_score):
   assignment_score_data = [name for name in grade_data.columns if assignment in name]
@@ -205,35 +187,6 @@ def create_course_eval_fig(course_eval_data, question, axes_labels):
   question_fig.for_each_annotation(lambda a: a.update(text=a.text[a.text.find("[")+1:a.text.find("]")]))
   return question_fig
 
-def create_sei_fig(sei_data: pd.DataFrame) -> plotly.graph_objs.Figure:
-  """
-  Creates an SEI data figure showing all of the SEI
-  data results over "time", where time is a categorical
-  semester string that is added to the SEI data. There
-  are four lines in this plot to compare against my
-  SEI data (i.e., the department, college, and university).
-  
-  :param sei_data: the raw SEI data as a dataframe
-  :return: the resulting SEI figure
-  """
-  sei_data["Semester"] = sei_data["Season"] + " " + sei_data["Year"].astype(str)
-  sei_fig = px.line(
-    sei_data, 
-    x="Semester", 
-    y="Mean", 
-    color="Group", 
-    facet_col="Question", 
-    facet_col_wrap=2, 
-    markers=True, 
-    title="Student Evaluation of Instruction Trends by Cohort",
-    category_orders={
-      "Semester": _semester_order(sei_data)
-    },
-    hover_data=["Course"]
-  )
-  sei_fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
-  return sei_fig
-
 def create_sei_comment_fig(sei_comments: pd.DataFrame) -> plotly.graph_objs.Figure:
   # Installs needed corpus data
   nltk.download('punkt')
@@ -319,14 +272,15 @@ def create_sei_tab() -> dcc.Tab:
         gap in teaching. 
         '''
       ),
-      dcc.Graph(id="bad-scale-1", figure=sei_fig),
+      dcc.Graph(id="sei-stats"),
       html.P(children=
         """
         Also, as a qualitative researcher, I find the comments themselves to be more meaningful.
         Therefore, here's a plot of the most frequent terms in my SEI comments. 
         """
       ),
-      dcc.Graph(figure=sei_comment_fig)
+      dcc.Graph(figure=sei_comment_fig),
+      load_sei_data()
     ])
 
 def create_course_eval_tab() -> dcc.Tab:
@@ -649,8 +603,6 @@ assignment_survey_data[review_count] = assignment_survey_data.groupby(project_re
 assignment_survey_data[std_time] = assignment_survey_data.groupby(project_review_col)[time_col].transform(lambda x: x.std())
 
 # SEI figures
-sei_data = pd.read_csv('https://raw.githubusercontent.com/jrg94/personal-data/main/education/sei-data.csv')
-sei_fig = create_sei_fig(sei_data)
 sei_comment_data = pd.read_csv('https://raw.githubusercontent.com/jrg94/personal-data/main/education/sei-comments.csv')
 sei_comment_fig = create_sei_comment_fig(sei_comment_data)
